PROJ-40 feat: add useNewPassword hook and refactor
