+++ features/newPassword/ui/NewPasswordForm.tsx
@@ -1,36 +1,21 @@
 'use client'
 
-import { FormProvider, useForm } from 'react-hook-form'
+import { FormProvider } from 'react-hook-form'
 
 import { ControlledInput } from '@/shared'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { Alert, Button, Card } from '@rambo-react/ui-meteors'
+import { Button, Card } from '@rambo-react/ui-meteors'
 import { clsx } from 'clsx'
-import { useRouter, useSearchParams } from 'next/navigation'
 
 import s from './NewPasswordForm.module.scss'
 
-import { useSetPasswordMutation } from '..'
-import { NewPasswordFields, newPasswordSchema } from '../model'
+import { useNewPassword } from '../hooks'
 
 export const NewPasswordForm = () => {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const recoveryCode = searchParams.get('recoveryCode')
-  const [setPassword, { error, isError, isLoading, isSuccess }] = useSetPasswordMutation()
+  const { isLoading, isRedirecting, methods, onSubmitHandler } = useNewPassword()
 
-  const methods = useForm<NewPasswordFields>({
-    mode: 'onBlur',
-    resolver: zodResolver(newPasswordSchema),
-  })
-
-  const onSubmitHandler = methods.handleSubmit(async (data: NewPasswordFields) => {
-    if (recoveryCode) {
-      await setPassword({ code: recoveryCode, password: data?.newPassword })
-        .unwrap()
-        .then(() => router.push('/auth/sign-in'))
-    }
-  })
+  if (isRedirecting) {
+    return null
+  }
 
   return (
     <Card className={s.card}>
@@ -51,13 +36,11 @@ export const NewPasswordForm = () => {
             type={'password'}
           />
           <p className={s.helper}>Your password must be between 6 and 20 characters</p>
-          <Button className={s.button} fullWidth type={'submit'}>
+          <Button className={s.button} disabled={isLoading} fullWidth type={'submit'}>
             Create new password
           </Button>
         </form>
       </FormProvider>
-      {!!error && <Alert message={(error as any).message} variant={'error'} />}
-      {!!isSuccess && <Alert message={'Password changed successfully'} variant={'accepted'} />}
     </Card>
   )
 }
